feat: add contact form to redux
