+++ src/components/Header.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Nav from './Nav';
+import Nav from './Nav/Nav';
 
 export default class Header extends Component {
   render() {
