+++ tests/pages/Forms.test.tsx
@@ -2,10 +2,22 @@ import { render, screen } from '@testing-library/react';
 import React from 'react';
 import userEvent from '@testing-library/user-event';
 import Forms from '../../src/pages/Forms';
+import { BrowserRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import store from '../../src/redux/store';
 
 describe('Forms Page:', () => {
+  beforeEach(() =>
+    render(
+      <Provider store={store}>
+        <BrowserRouter>
+          <Forms />
+        </BrowserRouter>
+      </Provider>
+    )
+  );
+
   it('Contains input', async () => {
-    render(<Forms />);
     const user = userEvent.setup();
     expect(screen.getByPlaceholderText("Enter contact's name and surname")).toBeInTheDocument();
     const input = screen.getByPlaceholderText("Enter contact's name and surname");
@@ -14,7 +26,6 @@ describe('Forms Page:', () => {
     expect(input).toHaveValue(text);
   });
   it('Try to fill the form incorrectly...', async () => {
-    render(<Forms />);
     const user = userEvent.setup();
     // expect(screen.getByPlaceholderText("Enter contact's name and surname")).toBeInTheDocument();
     const phoneNumber = 'someText1';
@@ -37,7 +48,6 @@ describe('Forms Page:', () => {
     ).toBeInTheDocument();
   });
   it('Try to fill the form correctly...', async () => {
-    render(<Forms />);
     const user = userEvent.setup();
     const phoneNumber = '+375292221453';
     const validName = 'Some Name';
@@ -54,5 +64,10 @@ describe('Forms Page:', () => {
     expect(screen.getByText('The data has been saved!')).toBeInTheDocument();
     const confirmButton = screen.getByText('Okay!');
     await user.click(confirmButton);
+    const defaultImg = screen.getByRole('img');
+    expect(defaultImg).toHaveAttribute(
+      'src',
+      'https://cdn-icons-png.flaticon.com/512/1077/1077114.png'
+    );
   });
 });
